ROM Type refactoring (OneUI)
Initial structure was too much based on the manufacturer, and not on the actual ROM type, which can differ for the same manufacturer.
Some are already split (MIUI/HyperOS), for others we need to split them in the future, when the next issue comes up (VIVO: FuntouchOS/OriginOS)

Renamed Samsung specific code to OneUI, as future Samsung AOSP ROMs might exist.
